Move CronPattern and CreateSchedule into a common client-accessible file
